requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
langchain-community
faiss-cpu
langchain-openai
tiktoken
langchain_google_genai